+++ opentelemetry-aws/src/trace/xray_propagator.rs
@@ -40,7 +40,7 @@
 
 use once_cell::sync::Lazy;
 use opentelemetry::{
-    global::{self, Error},
+    otel_error,
     propagation::{text_map_propagator::FieldIter, Extractor, Injector, TextMapPropagator},
     trace::{SpanContext, SpanId, TraceContextExt, TraceError, TraceFlags, TraceId, TraceState},
     Context,
@@ -139,7 +139,7 @@ pub fn span_context_from_str(value: &str) -> Option<SpanContext> {
             ))
         }
         Err(trace_state_err) => {
-            global::handle_error(Error::Trace(TraceError::Other(Box::new(trace_state_err))));
+            otel_error!(name: "SpanContextFromStr", error = format!("{:?}", TraceError::Other(Box::new(trace_state_err))));
             None //todo: assign an error type instead of using None
         }
     }